Issue #37. - Module "operationdialog" connecté avec module "composeur". - Bug à l'appel de la fonction "afterComposeurClose" réglé (voir commentaire ci-dessus). - Nettoyage effectué dans le module "operationDialog"
